Include a summary presenter for prisoner information
The idea here regarding the presenter is that we delegate method
calls to the underlying model, check if they are nil and return a
default text if they are or the value.

In cases where the underlying attribute needs to be presented in
a different manner we can override the underlying value and transform
the models data in the presenter class or return the default text.

Although the presenter  exposes all methods in the model class we
attempt not to allow mutation on the underlying model from the
presenters public interface as the method_missing does not allow
methods to pass any arguments.