Introduce summary page
Only includes prisoner information for now - this will be expanded to
include move information & a custody information overview for the
escort.

Makes use of the PrisonerInformationPresenter for displaying formatted
attributes pertaining to the person.

Slightly more work has been done in this commit than should be necessary
just to display the information currently available in the application
however it should serve as a guide for the next sections soon to arrive
in the app.