Simplify prisoner identification presenter
Removes the placeholder text("Not entered yet") and allows the return of
nil instead on properties that havent been filled in by the user
pertaining to a prisoners personal information such as names, age, sex
etc.

On the summary page the user should just see an empty space in the place
they would usually expect an attribute like forenames value. Its far
easier to see at a glance items that havent been filled in if they are
blank.