Use custom date parser
Introduces a custom date parser that allows Virtus to
correctly handle dates in various states that are acceptable
from a users point of view in the application.

We use a custom ParsedDate type that acts a wrapper around
the underlying Hash, Date or Nil value. It supports a consistent
interface for displaying the output on the form and for the date
object to be validated against.

The DateHashParser has the role of configuring the ParsedDate
correctly. ParsedDate offers a number of factory methods to
handle these actions.