Coerce & validate dates
Allows dates to be coerced correctly and provides a method
for displaying incorrect dates back to the user via a
presenter method.

Adds a class `date` method that handles configuring the
date attribute, validation, coercion & presentation on
the Virtus backed form objects.

This commit introduces translations to the form objects.