Add a text toggle virtus attribute helper
Provide a class method for introducing boolean & string
combinations as Virtus attributes in a form.

This will be useful for building out risk & health forms.
These forms require a boolean to accompany the textarea
that we persist so that we can determine if a user has
interacted with a section or not. This sort of data is
required so that we can tick yes, no or leave blank the
risk summary information both in the web app and on the
PDF.