Remove LastUpdated service
We have now broken the Escort record into multiple models, in order for
the auditing time stamp retrieval to work we will need to rethink the
approach. We are removing this code from the application so that we can
improve the approach & also because the code isn't required at the moment.