Add name, date of birth & last updated methods
Provide a way to extract a full name from a prisoner along with a
formatted date of birth.

Also adds a formatted last updated at method to the escort model.

This is necessary so that we can condense the logic down to one single
place(the model layer) rather than have it spread around the presentation
layer. This might have made sense earlier as we had one single object
that contained all of the data, we have since broken it up into different
models.