Filter request parameters in production mode
Replace all attributes(i.e. user input) sent to the server in
production mode so that they are not transferred over to the
ELK stack as the information is sensitive.

We can’t be sure that text in sections such as Risk doesn’t
contain information pertaining to an individual or a group
of individuals so having any sensitive data stored outside of
our application(i.e in ELK) is a bad idea.