Add PrisonNumber module
This module adds a prison_number attribute with field validation
on the format of value.

We use the same format as PVB1 & PVB2 projects to determine the
validity of a prison number. An example of a correct prison number
following this pattern would be: A1234BC.

This includes a shared example that can be included in any spec to
test the expected behaviour.