Test when an escort already exists
Provides a test for the rare case when a prison
number is no longer unique - for example it has
already been created in an alternate session by
another user.

At the moment we just redirect to the home page
when this happens.

This renames the search escort feature to account 
for the addition of this particular test. The home
page although providing search functionality is
also the point in which a user can create and seed
a new escort(with a prison number) or edit an
existing one.