Add User model
We are using devise & devise_invitable for
authentication & invite handling functionality
in the app.

The intended flow here is that an admin user in
the team will create a new prison user who will
receive an email allowing them to finalise their
account & setup a password.

The admin interface will only be accessible by the
MPS team at this stage.