Test inviting a user
We created an InvitationStep class to encapsulate 
the logic required to invite a user and set up a 
password to access the website. We need it because 
the raw_invitation_token, set on the user after the 
invite has been sent, is temporary and not accessible 
from subsequent steps.

Other than that, the spec tests the flow of a user 
accepting an invitation, clicking a link and setting
a password for the first time.