Extract urls from emails in feature specs
We extract password reset and invitation urls with tokens
in order to test devise functionality.

We removed the step pattern for the invitation feature spec.
Prior to this commit we held state across multiple steps in
order to have access to the devise token which was required
in order to hit the invitation accept end point provided by
devise.