Cache rendered output across assertions
Prior to this commit the render specs caused a massive
increase in spec runtime.

For now this ‘improvement’ caches the rendered output &
strips the html tags so we just have plain text in order
to improve the speed of ‘have_content’.

TODO: These tests need improving. Suggest transforming a
PDF back into html to & make assertions on individual
sections.