Avoid render call when testing pdf creation
Calls to rendering html are expensive, this commit
introduces a matcher that asserts a pdf file has
been created. The previous test didn’t actually assert
a PDF was created it just asserted that a PDFKit
instance was created.

These tests will more than likely change soon in order
to make assertions on the pdf content.