Provide an IE version for moj_template body_end
Change to address issue with partial IE conditional comment display
in the footer of the document in older IE. Identified as arising from
a change made to moj_template in:

https://github.com/ministryofjustice/moj_template/
commit/1d49dc033df5711469b137be067e5d2846d5a607

This change requires version to be set in the template. An addition
to the documentation for moj_template is being expored separately.