DRY form routes
Provide a central place for form names to be managed
for both routing and the navigation menu.

We rely on manually updating this list when forms are
added or removed in order to maintain a specific order
for the navigation menu, otherwise grepping the
ObjectSpace for form objects would’ve been a neater
solution.