Check guidance text exists for optional text areas
In order to make the toggle with details field more
generic we check if the translation for the guidance
field exists as this only exists on some fields. This
allows us to use the same partial for most of the 
healthcare attributes and to use the general details
text area with help text helper method on the
disabilities attribute which is quite different again
as it contains a checkbox.