Use repeated_permutation in healthcare spec
Sending a permutation message to the array has the action of
changing the arrangement, especially the linear order, of the
set of items(in this case [true, false nil] & in this spec
we allow repeats so it makes sense to use Arrays
repeated_permutation method as opposed to using product. While
we are at it lets change the method name so it matches with
what we are actually doing here.