Allow exclusion of descendentElement checkboxes
An interim measure while we re-evaluate our approach to generating
descendentElement checkboxes - that is those where visibility of
supplemental element(s) are dictated by a specific input state.

The previous implementation of descendentElements was based on:

1. There being only one set of checkboxes per block, each of which is
related to the descendent element

2. There being no 'nesting' of elements

During this sprint the requirement has changed and we will be changing
our approach to there generation of HTML for these elements. In light of
this it would be unwise to alter the HTML at this time and I have
therefore amended the JavaScript to allow developers to pass a number of
IDs to the plugin which will be ignored.

This interim measure is non-optimal and I expect the final code to be
substantially more succinct.