Introduce link_to_field_with_error helper
The link_to_field_with_error helper is needed to link
an error message to the field containing the error.

This helper manages the different types of fields with error.

If the field is a text toggle type, it will link to the name
of the field without _details termination.