Introduce a backfill collections service
Sometimes we need to prepend a number of empty objects to an
array(these might be empty models, forms, strings etc). This
commit introduces a way to build collections prepended with an
'empty thing' in a lazy manner. Given that an enumerator is
returned we can just `take` an amount off the top.