Provide dedicated form attributes transformer
As the forms have become more complex with the addition of
nested fields it is no longer straight forward to just pass
the forms `attributes`(hash) to active record. Some work has
to be done on the forms attributes in order for them to be
saved on the model correctly. We have decided to extract this
procedure out into a dedicated class that takes the current
forms attributes and returns a hash of attributes that will
be accepted by active record. We also use this as a place to
reject empties as well as in the case of nested models coerce
those into hashes too.