Accommodate long origin / destination names on PDF
MS noted this morning that layout for Handover section splits across
pages when a long origin or destination name is provided. Having
discussed the lack of space with JP it seems the most appropriate
response is to reduce the number of rows above to accommodate. Having
tested it seems removing a single row fixes the issue.